Big refactor : see description
Refactoring of data structure, we work on orderLines instead of Articles now.
Also, refactoring of some services methods (total is calculated and updated in cartService instead of having the same function everywhere.
The total is updated automatically when the cart changes or  when the Warehouse retrieval is (un)checked.

issues: ExpressionChangedAfterItHasBeenCheckedError appears when deleting all or one item in cart, also when taping the 'Réinitialiser' button in the article modal.